complete report menu, subclassed the bot
menu complete with 3 dropdowns and 2 buttons, 1 button for the description and proof, other for submitting and sending. persistent

this is done in a separate file;
made a new file for running the bot, now it's subclassed. This had to be done to allow the menu to persist through bot restarts;