datafetch, lexicon, watchdog
added a command to fetch users data and output the result as a csv.
Started working on moderation reversion. What i mean by that is a cog that listens for member updates made by mods(ban, timeout, ..) and pushes a poll on whether or not the decision should be reverted. Started assembling a lexicon to be used for the random word strings in report cards